+++ Space-Invaders/source/Gameplay/GameplayView.cpp
@@ -19,8 +19,8 @@ namespace Gameplay
 		sf::RenderWindow* game_window = ServiceLocator::getInstance()->getGraphicService()->getGameWindow();
 
 		background_image->initialize(Config::background_texture_path,
-			game_window->getSize().x,
-			game_window->getSize().y,
+			static_cast<float>(game_window->getSize().x),
+			static_cast<float>(game_window->getSize().y),
 			sf::Vector2f(0, 0));
 
 		background_image->setImageAlpha(background_alpha);
